Restore performance of image building on users' machines, plus other small fixes.
I've tried a few things,
but it turns out that running `uv sync` is quite IO-heavy,
and it's just not very fast if you don't have a warm disk cache.
So we have to stop doing that when building the image on the user's machine.

This does have the unfortunate consequence that we'll remember to update the base image whenever we update dependencies in pyproject.toml;
this shouldn't happen too frequently though.

Other small fixes:

- Install curl in Dockerfile.base:
  some subsequent commands expect it and are showing warning messages.

- Update uv to the latest 0.8.2.

- Remove the unnecessary `uv venv` command (the virtualenv will get created by `uv sync`).

- Write more in the README.md.